# Optimize StalinSort