Previous version used `.remove` method which was O(N).