@nvzqz 2019-11-09
I was intending on using tombstones to optimize this but this one-liner is really neat!