@nvzqz 2019-11-10
So this seems to not actually improve the performance of this implementation.

Before:

```
test stalin_sort::vec::benches::range_i032_1_0     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i032_1_1     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i032_1_10    ... bench:           8 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i032_1_100   ... bench:          85 ns/iter (+/- 1)
test stalin_sort::vec::benches::range_i032_1_1000  ... bench:         756 ns/iter (+/- 20)
test stalin_sort::vec::benches::range_i032_1_10000 ... bench:       7,511 ns/iter (+/- 266)
test stalin_sort::vec::benches::range_i064_1_0     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i064_1_1     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i064_1_10    ... bench:           8 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i064_1_100   ... bench:          84 ns/iter (+/- 2)
test stalin_sort::vec::benches::range_i064_1_1000  ... bench:         759 ns/iter (+/- 30)
test stalin_sort::vec::benches::range_i064_1_10000 ... bench:       7,500 ns/iter (+/- 177)
test stalin_sort::vec::benches::range_i128_1_0     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i128_1_1     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i128_1_10    ... bench:          14 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i128_1_100   ... bench:         157 ns/iter (+/- 4)
test stalin_sort::vec::benches::range_i128_1_1000  ... bench:       1,503 ns/iter (+/- 32)
test stalin_sort::vec::benches::range_i128_1_10000 ... bench:      15,001 ns/iter (+/- 920)
```

After:

```
test stalin_sort::vec::benches::range_i032_1_0     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i032_1_1     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i032_1_10    ... bench:           8 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i032_1_100   ... bench:          86 ns/iter (+/- 9)
test stalin_sort::vec::benches::range_i032_1_1000  ... bench:         765 ns/iter (+/- 40)
test stalin_sort::vec::benches::range_i032_1_10000 ... bench:       7,632 ns/iter (+/- 509)
test stalin_sort::vec::benches::range_i064_1_0     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i064_1_1     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i064_1_10    ... bench:           8 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i064_1_100   ... bench:          86 ns/iter (+/- 6)
test stalin_sort::vec::benches::range_i064_1_1000  ... bench:         760 ns/iter (+/- 22)
test stalin_sort::vec::benches::range_i064_1_10000 ... bench:       7,520 ns/iter (+/- 379)
test stalin_sort::vec::benches::range_i128_1_0     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i128_1_1     ... bench:           1 ns/iter (+/- 0)
test stalin_sort::vec::benches::range_i128_1_10    ... bench:          15 ns/iter (+/- 1)
test stalin_sort::vec::benches::range_i128_1_100   ... bench:         156 ns/iter (+/- 3)
test stalin_sort::vec::benches::range_i128_1_1000  ... bench:       1,502 ns/iter (+/- 60)
test stalin_sort::vec::benches::range_i128_1_10000 ... bench:      15,086 ns/iter (+/- 1,787)
```

However, since it's not a regression, I'll merge this since it's much cleaner than what was there before. I guess I'll have to go with tombstones anyway.