@xfix 2019-11-10
To be fair, the benchmarks in your example were testing the best case performance of previous algorithm when nothing got removed.

```rust
let mut vec: Vec<$i> = (1..=$max).collect();
```

Feel free to compare this with this:

```rust
let mut vec: Vec<$i> = (1..=$max).rev().collect();
```